+++ src/greptimeai/langchain_tracker/callback.py
@@ -14,9 +14,9 @@ from langchain.schema.output import (
 )
 from tenacity import RetryCallState
 
-from greptimeai import logger
 from greptimeai.collection import Collector
-from greptimeai.openai.utils.tokens import (
+from greptimeai.logger import logger
+from greptimeai.openai_tracker.utils.tokens import (
     cal_openai_token_cost_for_model,
     num_tokens_from_messages,
 )
