+++ src/greptimeai/logger.py
@@ -1,0 +1,5 @@
+import logging
+
+from .scope import _NAME, _VERSION
+
+logger = logging.getLogger(f"{_NAME}:{_VERSION}")
