+++ tests/openai_tracker/test_tokens.py
@@ -1,4 +1,4 @@
-from greptimeai.openai.utils.tokens import (
+from greptimeai.openai_tracker.utils.tokens import (
     cal_openai_token_cost_for_model,
     num_tokens_from_messages,
 )
